[ASV-455] Removed MyAccountFragment and MyAccountNavigator
