+++ app/src/vanilla/java/cm/aptoide/pt/view/settings/NewAccountNavigator.java
@@ -1,15 +1,18 @@
 package cm.aptoide.pt.view.settings;
 
+import android.app.Activity;
 import cm.aptoide.pt.account.AccountAnalytics;
 import cm.aptoide.pt.account.view.AccountNavigator;
-import cm.aptoide.pt.account.view.MyAccountNavigator;
 import cm.aptoide.pt.account.view.store.ManageStoreFragment;
 import cm.aptoide.pt.account.view.store.ManageStoreViewModel;
+import cm.aptoide.pt.account.view.user.ManageUserFragment;
 import cm.aptoide.pt.addressbook.view.AddressBookFragment;
 import cm.aptoide.pt.app.AppNavigator;
 import cm.aptoide.pt.dataprovider.model.v7.store.Store;
 import cm.aptoide.pt.navigator.FragmentNavigator;
 import cm.aptoide.pt.notification.view.InboxFragment;
+import cm.aptoide.pt.store.StoreTheme;
+import cm.aptoide.pt.store.view.StoreFragment;
 import rx.Observable;
 
 /**
@@ -19,17 +22,15 @@ import rx.Observable;
 public class NewAccountNavigator {
 
   private final FragmentNavigator fragmentNavigator;
-  private final MyAccountNavigator accountNavigator;
   private final AccountNavigator navigator;
   private final AppNavigator appNavigator;
 
   private final String UPLOADER_UNAME = "aptoide-uploader";
   private final String BACKUP_APPS_UNAME = "aptoide-backup-apps";
 
-  public NewAccountNavigator(FragmentNavigator fragmentNavigator,
-      MyAccountNavigator accountNavigator, AccountNavigator navigator, AppNavigator appNavigator) {
+  public NewAccountNavigator(FragmentNavigator fragmentNavigator, AccountNavigator navigator,
+      AppNavigator appNavigator) {
     this.fragmentNavigator = fragmentNavigator;
-    this.accountNavigator = accountNavigator;
     this.navigator = navigator;
     this.appNavigator = appNavigator;
   }
@@ -47,23 +48,31 @@ public class NewAccountNavigator {
   }
 
   public void navigateToEditStoreView(Store store, int requestCode) {
-    accountNavigator.navigateToEditStoreView(store, requestCode);
+    ManageStoreViewModel viewModel = new ManageStoreViewModel(store.getId(), StoreTheme.fromName(
+        store.getAppearance()
+            .getTheme()), store.getName(), store.getAppearance()
+        .getDescription(), store.getAvatar(), store.getSocialChannels());
+    fragmentNavigator.navigateForResult(ManageStoreFragment.newInstance(viewModel, false),
+        requestCode, true);
   }
 
   public Observable<Boolean> editStoreResult(int requestCode) {
-    return accountNavigator.editStoreResult(requestCode);
+    return fragmentNavigator.results(requestCode)
+        .map(result -> result.getResultCode() == Activity.RESULT_OK);
   }
 
   public void navigateToEditProfileView() {
-    accountNavigator.navigateToEditProfileView();
+    fragmentNavigator.navigateTo(ManageUserFragment.newInstanceToEdit(), true);
   }
 
-  public void navigateToUserView(String userId, String theme) {
-    accountNavigator.navigateToUserView(userId, theme);
+  public void navigateToUserView(String userId, String storeTheme) {
+    fragmentNavigator.navigateTo(
+        StoreFragment.newInstance(userId, storeTheme, StoreFragment.OpenType.GetHome), true);
   }
 
-  public void navigateToStoreView(String storeName, String theme) {
-    accountNavigator.navigateToStoreView(storeName, theme);
+  public void navigateToStoreView(String storeName, String storeTheme) {
+    fragmentNavigator.navigateTo(
+        StoreFragment.newInstance(storeName, storeTheme, StoreFragment.OpenType.GetStore), true);
   }
 
   public void navigateToLoginView(AccountAnalytics.AccountOrigins accountOrigins) {
